Updated sql in order to run on server

+++ conf/evolutions/default/2.sql
@@ -10,7 +10,7 @@ CREATE TABLE oppmerksomhet (
     fraType varchar(32),
     url varchar(255),
     info varchar(255),
-    verdi double,
+    verdi double PRECISION,
     tid timestamp,
     hendelsestype varchar(32),
     levert boolean,
